\name{cv.np}
\alias{cv.np}
\title{
Calculate prodiction MSE for nonparametric method
}
\description{
Function to calculate prodiction MSE for nonparametric method.
}
\usage{
cv.np(Y, S, A, Ytest, Stest, Atest)
}
\arguments{
  \item{Y}{
}
  \item{S}{
}
  \item{A}{
}
  \item{Ytest}{
}
  \item{Stest}{
}
  \item{Atest}{
}
}

\value{
a numeric value; MSE in cross-validation.
}
\references{
Duan, Y. and Parast, L., 2023. Flexible evaluation of surrogate markers with Bayesian model averaging. Statistics in Medicine.
}
\author{
Yunshan Duan
}

\keyword{internal}
