\name{exampleData}
\alias{exampleData}
\docType{data}
\title{
Example data
}
\description{
Example data
}
\usage{data("exampleData")}
\format{
 A data frame with 400 rows and 3 variables:
  \describe{
  	  \item{\code{Y}}{the primary outcome}
  	  \item{\code{S}}{the surrogate marker}
  	  \item{\code{A}}{the treatmentt arm}
}
}
\examples{
data(exampleData)
names(exampleData)
}
