\name{fpred.int}
\alias{fpred.int}
\title{
Function to be integrated when calculating the expected primary outcome
}
\description{
Function to be integrated when calculating the expected primary outcome.
}
\usage{
fpred.int(eps, mu, sig, m)
}
\arguments{
  \item{eps}{
}
  \item{mu}{
}
  \item{sig}{
}
  \item{m}{
}
}

\value{
a function.
}
\references{
Duan, Y. and Parast, L., 2023. Flexible evaluation of surrogate markers with Bayesian model averaging. Statistics in Medicine.
}
\author{
Yunshan Duan
}


\keyword{internal}
