\name{pred.func}
\alias{pred.func}
\title{
Calculate the expected primary outcome in the treatment group given the model and the parameters
}
\description{
Provides a function to calculate the expected primary outcome in the treatment group given the model and the parameters.
}
\usage{
pred.func(theta, ST, m)
}
\arguments{
  \item{theta}{
}
  \item{ST}{
}
  \item{m}{
}
}
\value{
a numeric vector.
}
\references{
Duan, Y. and Parast, L., 2023. Flexible evaluation of surrogate markers with Bayesian model averaging. Statistics in Medicine.
}
\author{
Yunshan Duan
}


\keyword{internal}
