\name{R.m.theta}
\alias{R.m.theta}
\title{
Calculates the R value given model and parameters
}
\description{
Provides a function to calculate F_m(theta_m).
}
\usage{
R.m.theta(S, m, theta, A, weight = FALSE, nBB = 100)
}
\arguments{
  \item{S}{
}
  \item{m}{
}
  \item{theta}{
}
  \item{A}{
}
  \item{weight}{
}
  \item{nBB}{
}
}
\value{
a numeric value or a numeric verctor when weight = TRUE. the value of R or the values of R using Bayesian bootstap replicates of the surrogate marker.
}

\author{
Yunshan Duan
}


\keyword{internal}
