\name{SurrogateBMA-package}
\alias{SurrogateBMA-package}
\alias{SurrogateBMA}
\docType{package}
\title{
  Flexible Evaluation of Surrogate Markers with Bayesian Model Averaging
}
\description{
Provides functions to estimate the proportion of treatment effect explained by the surrogate marker using Bayesian Model Averaging approach.
}
%% \details{
  %% This section should provide a more detailed overview of how to use the
  %% package, including the most important functions.
%% }
\author{
Yunshan Duan

Maintainer: Yunshan Duan <yunshan@utexas.edu>
}
\references{
  Duan and Parast (2023+). Flexible Evaluation of Surrogate Markers with Bayesian Model Averaging.
}
\keyword{ package }
%% \seealso{
  %% Optional links to other man pages
%% }
%% \examples{
  %% \dontrun{
     %% ## Optional simple examples of the most important functions
     %% ## These can be in \dontrun{} and \donttest{} blocks.   
  %% }
%% }

\keyword{internal}
