\name{SurrogateBMA-package}
\alias{SurrogateBMA-package}
\alias{SurrogateBMA}
\docType{package}
\title{
  Flexible Evaluation of Surrogate Markers with Bayesian Model Averaging
}
\description{
Provides functions to estimate the proportion of treatment effect explained by the surrogate marker using Bayesian Model Averaging approach.
}
\author{
Yunshan Duan

Maintainer: Yunshan Duan <yunshan@utexas.edu>
}
\references{
Duan, Y. and Parast, L., 2023. Flexible evaluation of surrogate markers with Bayesian model averaging. Statistics in Medicine.
}
\keyword{ package }


\keyword{internal}
